Handle the PATH
Giving Any command the shell is able to find the path using the custom
pathfinder(), which helps executor(), do decide execution if command is
present.

This gives the shell the ability to run any command present on the sys.
Rather than having to stress the user to input the absolute path to it.
Though there are side effect, since the shell has to make more syscalls.

Issues yet to be solved:

 - There still exist library calls

 - Error mssgs specifying how a cmds should be used displays abs path

 - Error messages are not appropiate for normal users.